Import changes for goma client
  - 8997be122059586f749fbc6438c865484b375211 Follow up CL to fix build issue.
  - e4fea769592f0767dcff9e7e13fda38842f2d38b Make build/*.py compatible with pylint-2.6.
  - 0a1035caffae8cfa94ad91501eb65a00bcf89552 Pass FORCE_DISABLE_WERROR environment
  - 5892d6a5f5c39c469c116b82a98fc2c9f73fb487 Add autoroller service account in OWNERS
  - 13ceb1a36acbb741578aa5ba6a6f307e3f51b53f Roll clang from ac0a49ca33a6 to 5ab37cd7bdc3 (1 revision)
  - 50797da4f31f972f2d8020d12381e3df98b4932b Fix NaClCompilerInfoBuilderHelper::CollectNaClClangResource
  - 746c0e76fa8db3bcc162dd3c9f56dbccea3e0658 Roll boringssl from d55f450c4fa6 to c6d3fd1d0972 (1 revis...
  - ace0d545baeb6d7cdc50224468c89e07555a12b8 Remove .vpython3.
  - 1199c06eab2f37a371390311bb41d9f2eb610716 Roll boringssl from a75027b0401c to d55f450c4fa6 (1 revis...
  - 7367304867163b432afa3109d97d3db713754ad3 Roll boringssl from e2cb423767cc to a75027b0401c (1 revis...
  - 7b3c103b1e9351cfecbf384fb8ca8995a25da491 Roll boringssl from 61f320874be3 to e2cb423767cc (1 revis...
  - b1842861cb45cfd14cdb72c92c89f9abbe845f57 Roll clang from 3f6acf9843fb to ac0a49ca33a6 (1 revision)
  - 84fccef49913f47a7c2add7309d68bad9a43bd01 Roll clang from cb4a58d41460 to 3f6acf9843fb (2 revisions)
  - 5dd564f613243fb2d03549237b470b6ae4780121 Roll boringssl from 4bf0a19ac843 to 61f320874be3 (3 revis...
  - 7dc8eb4d8621753487a76941a2ab018f6cc69859 Roll boringssl from 047ff6428714 to 4bf0a19ac843 (7 revis...
  - da84c4c8092b1b0c685e85c59e600aa4264ff4da Make gn scripts run with python3.
  - 9f31575f4838e806a795e1131b7a98244a967fdc Fix NaClCompilerInfoBuilderHelper::CollectNaClClangResour...
  - 4a9b914231a4851bd83a76fa63e6dad3b8e86e14 Updated minimum macOS version.
  - ea2ea371d0b6b260bc314ee9b1e9728014d11dd4 Make dump_syms.py works with Python3.
  - 12c629585c175a1b8c620b8067185e5554df1fdf cleanup: make mac_sdk.gni depends on third_party/chromium...
  - 0f82cf585dde33d78aeecc52dd3de7b932cbdbab Roll clang from 6002926a08a5 to cb4a58d41460 (2 revisions)
  - 3ad4e3a204109a595af1ebe1f3f607dfa92e0850 cleanup: use mac python scripts in chromium_build.
  - ccb8337bac166ddc230481bcbfb9a04506e18ee0 Update eary-access-guide.md
  - 9c8b5d8c20b063f899271b3ee9f3ea6c205bfcb1 Update go to 1.17
  - 609e6655143e381827d4cb90a14e6358eb810943 Roll boringssl from 8648c5369020 to 047ff6428714 (1 revis...
  - e713481eca406c6cda102e489197cf7ae1718a29 Roll boringssl from 16c3e3ae0e60 to 8648c5369020 (1 revis...
  - bd55ac7a00ee5a895667d85adeeae27e294d1c39 Roll boringssl from 05ce773cae19 to 16c3e3ae0e60 (1 revis...
  - a44480ff8308c850a002169d16424e596c9c4814 Roll boringssl from 80df7398ce52 to 05ce773cae19 (1 revis...
  - aa02ff0f5afdee0f5932555fbc22661cc026dd94 fix simpletry.sh on mac.
  - 600be08779608fa131dc005dae9759bb455f3cab Roll boringssl from a603c828dc29 to 80df7398ce52 (1 revis...
  - 64d35858da5054db825f0769c1ecd0ac937d0ddc Roll boringssl from 006f20ad7f9a to a603c828dc29 (1 revis...
  - a605a30128dcc77c2cf2bbd62f1e2fae7aca1ffe use python3 to run "gclient runhook"
  - 6b47c7db9a85ac8ddee63f8bc12f7cef0f7458f2 Roll clang from f1ab49ba2281 to 6002926a08a5 (1 revision)
  - d06fc37e7221ebe50da09df0b1320919a6fb64f1 cleanup: remove build/mac*.py
  - dfc0dc21565c32c958ea32fe0215ac4e4052aef5 cleanup: remove lastchange.*.
  - cfdb9f62917122bd51652850eb6516b27bb52ea4 Update build/util/lastchange.* to commit f0822118712d
  - 5f3acc0c2d217ce3972c99dc2d1e10cc1ecb1a1d cleanup: remove test/third_party/binutils
  - 85ead967f4e798e580ecffdfeefae0d49b61ac8e Roll boringssl from 2e68a05c9943 to 006f20ad7f9a (1 revis...
  - 6046ded0c1a335e7953a2d938372f88dee22a796 Roll client/third_party/chromium_build/ 909847cd6..cb0fa2...
  - 4147be65b8af9dd0b70673a0b2093a5459005183 generate_compiler_proxy_info: stop using win32api
  - eda4841cc44495d201f8058a0f86cd2977728fe6 Roll boringssl from 69ec7c8de73e to 2e68a05c9943 (1 revis...
  - 9177ce202b66153b8d8f2c0748ba953a537a6983 Roll boringssl from b9ee7b143135 to 69ec7c8de73e (1 revis...
  - 9379bb450ad980ff02588382e8bab727969a7ff5 Detach from tty when creating detached processes.
  - d0bbc26d4238ba2a5dc5369ea4e52fade559e0a2 Roll boringssl from eb17de499a36 to b9ee7b143135 (2 revis...
  - 3877907e981abe0d8d33fefa0223a6bbd0722835 Roll clang from 6a8e571efd68 to f1ab49ba2281 (1 revision)
  - 97de0923b1a8e554791d9eb6c9f13cc6e7f7d481 Roll boringssl from e3a365554293 to eb17de499a36 (2 revis...
  - 57fae5f2bc70637c0903b900f23efdd3ae7a6d21 Roll boringssl from 4c993da66019 to e3a365554293 (1 revis...
  - 2869a04cf2090b46a296c4802c0dda82ff9f1562 Roll boringssl from 07a6628e4b1b to 4c993da66019 (1 revis...
  - 0ed610b68c0b6d60fe720fe562a2227718741f4b Roll client/buildtools/ 69cc9b8a3..37dc929ec (31 commits;...
  - bdbfb9b26901af2928c1445c715d3604a5f4a8f3 cleanup: remove python2 compatiblity from code run by bots.
  - 60908d7ebf739ce85acb36b68dd3155edf368853 Roll boringssl from 47c5f9d2f6c2 to 07a6628e4b1b (10 revi...
  - 5cb12a8733a9856006abdc8fa6b81444282dd4f6 make get_diff_from_previous and simpletry run with python3.
  - e1ce413228f1ce6744ad462a35a7488aa3d2856b Unlimit the deps cache table size.
  - fc261ab5830a06620afbc8c8f43fd2f64c166f65 Treat PNACLPYTHON as important env.
  - c0650132df1798570b586aba34ac086016336d53 Unlimit deps_cache file size by default.
  - 3d0cc2809e9959f4aa9fbaac9e68446d1699d94d Roll boringssl from 549e4e7995a9 to 47c5f9d2f6c2 (1 revis...
  - 1b37949e5b67cfe0dcc27526bcfb8a7bee29cacd Roll boringssl from 7e265971c0b9 to 549e4e7995a9 (1 revis...
  - 51932c7a563578f334653a1ebf3aea6c104f2162 Roll clang from a6e918faf005 to 6a8e571efd68 (1 revision)
  - 78140e217efacb17e3fc563e506597eae4be3317 Roll clang from dd02422d012e to a6e918faf005 (1 revision)
  - ca6b488be164feafea4f96ee9f16e5aaf69a892a Roll boringssl from ead57c3004fe to 7e265971c0b9 (1 revis...
  - 3946d786ca2c759b77cb30a39054f6cbc95e0c61 Roll boringssl from 046fc130d1cd to ead57c3004fe (2 revis...
  - 2d656c64d5db2e9a83994fd0c3efc84c4a9c59d6 Roll boringssl from 116d9250a996 to 046fc130d1cd (1 revis...
  - 346e920b2f664ce900d7ffdef764e10a9f1a8c4d Roll boringssl from e9fae77c06df to 116d9250a996 (1 revis...
  - 89505c2bb1081a3982ff15ae776ab24d7fb2880b Roll boringssl from 0768d42c283d to e9fae77c06df (1 revis...
  - 20a864c0fd312f975532e72f927d355df733e628 Roll boringssl from f1d153dc3619 to 0768d42c283d (1 revis...
  - 93d30fcc9da72aee2eb0f4f87d153477f6b82f2a Roll boringssl from 5799ebfe5f89 to f1d153dc3619 (1 revis...
  - 2ad1bccfb6c4da6ef28e0b9c1735970c0a06dbf4 Roll clang from d0097bccbaab to dd02422d012e (1 revision)
  - c8567ae009479f9e962b0edbbe8dffa8eb9f8e03 Roll clang from 4b88fb756103 to d0097bccbaab (1 revision)
  - 13f185e43a357a96c104378d685b1476ed5bfb02 Roll clang from 976f00e339a4 to 4b88fb756103 (1 revision)
  - 1f740f06bea7d85cf77d6b98bd859e996ddf0cff Roll boringssl from d422d2c4aada to 5799ebfe5f89 (1 revis...
  - 03bd4c183a5e4f45481b57cf1d8a33154a745a94 Roll clang from 30a89ee609ec to 976f00e339a4 (1 revision)
  - 1b8fb8308e259fbb01b5bccaf48de1a544c02ca5 Roll clang from 3d34ce6bf562 to 30a89ee609ec (1 revision)
  - ba32b96b7704a18116022abc0fd3088fd5cedb06 Roll clang from d76ac6af4b0b to 3d34ce6bf562 (2 revisions)
  - c8dee742486471937c33651a5178bd3a68b33de6 Roll clang from 2a8bb1ca80b1 to d76ac6af4b0b (1 revision)
  - c421480b1fa0298522f6b1236a20bb4a35e3f297 Roll clang from 94ebf0661eb5 to 2a8bb1ca80b1 (4 revisions)
  - 09544adaf8864ace8cd1888f907bfbd254988a25 Roll boringssl from c1571feb5faf to d422d2c4aada (1 revis...
  - 5de13e597ee3ad2939167729aea4d8df80b34e1d Roll clang from 131233f78952 to 94ebf0661eb5 (1 revision)
  - 7be9b042e7da7b440c7b98f3dfe35bfdd8cfd19f Roll boringssl from 7a817f48bafe to c1571feb5faf (1 revis...
  - 0ee6e02b7f5f34e36754e4c308b85c73d1d31113 Roll clang from 843b4269cfbc to 131233f78952 (1 revision)

GitOrigin-RevId: 8997be122059586f749fbc6438c865484b375211

+++ client/deps_cache.h
@@ -51,7 +51,7 @@ class DepsCache {
   // When |cache_filename| file exists, call LoadIfEnabled to load cache file.
   static void Init(const std::string& cache_filename,
                    absl::optional<absl::Duration> identifier_alive_duration,
-                   size_t deps_table_size_threshold,
+                   int deps_table_size_threshold,
                    int max_proto_size_in_mega_bytes);
   // Load cached data from cache_filename,
   // when cache_filename is not empty.
@@ -132,7 +132,7 @@ class DepsCache {
 
   DepsCache(const std::string& cache_filename,
             absl::optional<absl::Duration> identifier_alive_duration,
-            size_t deps_table_size_threshold,
+            int deps_table_size_threshold,
             int max_proto_size_in_mega_bytes);
   ~DepsCache();
 
@@ -169,7 +169,7 @@ class DepsCache {
   const absl::optional<absl::Duration> identifier_alive_duration_;
   // When lots of DepsTable exist, we'd like to remove older DepsTable entry
   // when saving.
-  const size_t deps_table_size_threshold_;
+  const int deps_table_size_threshold_;
   // If the proto for cache exceeds this size, loading will fail.
   // In that case, cache is just ignored.
   const int max_proto_size_in_mega_bytes_;
